Moved model data into its own class.
